serialio: Refactor serial/ from util/
