+++ kernel/makefile
@@ -5,9 +5,9 @@ BUILD_DIR     = ../build/kernel
 TARGET        = $(BUILD_DIR)/kernel.bin
 ASM_OBJS      = kernel_asm.o util/portio_asm.o
 OBJS          = kernel.o video/video.o interrupt/interrupt.o \
-                interrupt/handler.o util/serial.o
+                interrupt/handler.o serial/serial.o
 INCLUDE       = util/portio.h video/video.h util/types.h interrupt/interrupt.h \
-                util/keyboard_map.h config.h util/serial.h
+                util/keyboard_map.h config.h serial/serial.h
 
 all:          $(TARGET)
